crud user -v0.1

+++ src/modules/user/user.service.ts
@@ -1,10 +1,16 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  HttpException,
+  Injectable,
+  Logger,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from './../prisma/prisma.service';
 import { User, Prisma } from '@prisma/client';
 
 @Injectable()
 export class UserService {
-  constructor(private prisma: PrismaService) {}
+  constructor(private prisma: PrismaService, private readonly logger: Logger) {}
 
   async user(
     userWhereUniqueInput: Prisma.UserWhereUniqueInput,
@@ -31,26 +37,56 @@ export class UserService {
     });
   }
 
-  async createUser(data: Prisma.UserCreateInput): Promise<User> {
-    return this.prisma.user.create({
-      data,
-    });
+  async createUser(data: Prisma.UserCreateInput): Promise<any> {
+    try {
+      const { username, password, fullName } = data;
+      const existedUser = await this.user({ username });
+
+      if (existedUser) throw new Error('Username already exist');
+
+      return this.prisma.user.create({
+        data: { username, password, fullName },
+      });
+    } catch (error) {
+      this.logger.error(`${error.message}`);
+      throw new BadRequestException(error.message);
+    }
   }
 
   async updateUser(params: {
     where: Prisma.UserWhereUniqueInput;
     data: Prisma.UserUpdateInput;
   }): Promise<User> {
-    const { where, data } = params;
-    return this.prisma.user.update({
-      data,
-      where,
-    });
+    try {
+      const { where, data } = params;
+      const existedUser = await this.user({ id: where.id });
+
+      if (!existedUser) throw new Error('Accout Not Found');
+
+      return this.prisma.user.update({
+        data,
+        where,
+      });
+    } catch (error) {
+      this.logger.error(`${error.message}`);
+      throw new BadRequestException(error.message);
+    }
   }
 
-  async deleteUser(where: Prisma.UserWhereUniqueInput): Promise<User> {
-    return this.prisma.user.delete({
-      where,
-    });
+  async deleteUser(where: Prisma.UserWhereUniqueInput): Promise<any> {
+    try {
+      const existedUser = await this.user(where);
+
+      if (!existedUser) throw new Error('Accout Not Found');
+
+      await this.prisma.user.delete({
+        where,
+      });
+
+      return true;
+    } catch (error) {
+      this.logger.error(`${error.message}`);
+      throw new BadRequestException(error.message);
+    }
   }
 }
